backend/requirements.txt
mysqlclient==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
Django==4.2.5
requests==2.31.0
numpy==1.26.0
whitenoise==6.5.0
bcrypt==4.0.1
pyyaml==6.0.1
uritemplate==4.1.1
